backend/mt5/app/requirements.txt
numpy==1.22.4 
pandas==1.4.2
pytz
python-dotenv
flasgger
python-json-logger
flask
MetaTrader5
PyJWT==2.8.0